# Summary
Upstream recent changes to fMHA that we did in xFormers.
Current version in CUTLASS: https://github.com/facebookresearch/xformers/commit/b6be33aecb5297f3f994568cf29e194a75e47667
Updating to: https://github.com/facebookresearch/xformers/commit/55a4798d99c350eca9d225a14003f226775e5f76

# Changelog
Full changelog: https://github.com/facebookresearch/xformers/commits/main/xformers/csrc/attention/cuda
**FW pass:**
* Change how we do the iterative softmax to reduce the number of operations + reduce calls to `__syncthreads()` (https://github.com/facebookresearch/xformers/commit/26884142271d5fbf897c69cc43dd0523e8f9f87a)
* Remove the "DISPATCH_BOOL" inside the kernel - this was inducing a lot of RF use for some reason. The idea was to check only once if we could skip all bounds checks (to avoid checking bounds multiple times) but apparently it's making things worse (https://github.com/facebookresearch/xformers/commit/26884142271d5fbf897c69cc43dd0523e8f9f87a)
* Stored `warp_id` in a RF and reuse it, rather than fetching it from `threadIdx.x` multiple times (https://github.com/facebookresearch/xformers/commit/26884142271d5fbf897c69cc43dd0523e8f9f87a)
* Make kernel deterministic (no longer using atomicAdds) (https://github.com/facebookresearch/xformers/commit/b63f634765414b03f5310adc140fa49eaeb2e6e7)
* The kernel now takes an upperbound on the maximum reduction dimension supported: this allows some further optimizations if we know that the operand fits entirely in shmem, as we no longer have to maintain pointers to where we write to shmem (as everything is written during prologue) (https://github.com/facebookresearch/xformers/commit/73245e69f80b64521f0eefebf5bf6fa1cde2fc40)
* We recommend to increase blockSize to 64x128 rather than 32x128 after Sm75 (https://github.com/facebookresearch/xformers/commit/7dbcc363e81260a566efc2e87a3bb1a614beb726)
* Fix overflow on sequence length (https://github.com/facebookresearch/xformers/commit/68dce697013d3078315174840106bc9a81477d82#diff-dddb484afe22b9bc99f6c0d0f66d21d4cff0858b22dd7abeb37dcb63309ab9a9)
* BUGFIX when using attention bias with `-inf` prefix that exceeds the block size (https://github.com/facebookresearch/xformers/commit/540fcbfdaf8230fd37644345a310a36af86c9d6b)
* BUGFIX: Overflow in bias stride for very long sequences (https://github.com/facebookresearch/xformers/commit/471569b4c8adec77ba79023cfa9e9f5598284542)
* Fix race condition with `cp.async` copies still being in flight after we have left the MMA inner loop (https://github.com/facebookresearch/xformers/commit/55a4798d99c350eca9d225a14003f226775e5f76)

**BW pass:**
* Fix dropout seed when using variable sequence lengths (aka packed tensors) (https://github.com/facebookresearch/xformers/commit/70161e5592a116801f1ba588f7797a3b1becb0f3)
* Multiple performance improvements https://github.com/facebookresearch/xformers/commit/5e5237324b8168c42f7449702a41220a700b2666
* Kernel accepts a template argument to allow it to skip bound checks entirely - if the problem size is aligned with the block sizes (https://github.com/facebookresearch/xformers/commit/7f137800b392070f96b56511f63f7b0360cd0161)
* Added a few tricks to reduce register pressure (https://github.com/facebookresearch/xformers/commit/936da0a21e7ee3997c262f42d79b7e26ad4a2f69 https://github.com/facebookresearch/xformers/commit/f1a0d9cd73ae22be56b0ea4cfc8f2d1d05de19ea https://github.com/facebookresearch/xformers/commit/a122eae0339dd454bac8614021b6f291e655fe6b https://github.com/facebookresearch/xformers/commit/f2aa13a659bd028c0f76da3addcabfdd53ed109d https://github.com/facebookresearch/xformers/commit/e2b9ec5fbd4b32949f9ccb69afa52a4762d078c0) - this massively improves performance on CUDA 11.7+
* Parallelize on sequence length - this gives massive speedups when `batch_size * sequence_length << number_of_sms * parallelization` (NOTE: This makes the kernel not deterministic when enabled) (https://github.com/facebookresearch/xformers/commit/19902ae0a4a60a484a4938cb0c4fa762e339fe84)
* Fix race condition (https://github.com/facebookresearch/xformers/commit/5498468a3555aab24703b9aa569f17f945397af3)

## New performance options to tune:
* **BW & FW**: Set `kMaxK` to an upperbound of the embedding per head (16, 32, 64, 96, 128...)
* **BW**: Paralellization across keys: set `num_splits_key` to improve parallelization at the cost of more memory usage. Might require some tuning (see the [heuristic](https://github.com/facebookresearch/xformers/blob/55a4798d99c350eca9d225a14003f226775e5f76/xformers/csrc/attention/cuda/fmha/attention_backward_generic.cu#L311-L334) we use in xformers as an example). If not using it, you can disable it in the template parameters for better performance (`kEnableSplitKeys=false`)
* **BW**: To skip bounds checks, you can also enable `kKeysQueriesAlignedToBlockSize = true` when your problem size is aligned with block size (eg block size = 64x128 and sequence length=1024)

# Performance benchmarks
* `eager`/`vanilla` = pytorch eager mode (when empty, it means OOM)
* `v2` = current version in CUTLASS
* `v3` = this PR
* Shape is in format `batch_size-seq_len-num_head-dim_per_head`

**A100 SMX 80GB** (cuda 11.8)
<details>
<summary>FW - up to 20% faster</summary>

```
[--------------------- attention (attn_bias=<class 'NoneType'>) ---------------------]
                                                 |     v3     |     v2     |   eager  
1 threads: ---------------------------------------------------------------------------
      f16 384-197-1-88, p=0.0, BiasT=NoneType    |     132.2  |     138.0  |     416.0
      f16 384-197-1-80, p=0.0, BiasT=NoneType    |     126.4  |     129.0  |     339.0
      f16 384-197-1-64, p=0.0, BiasT=NoneType    |      83.1  |      94.0  |     289.0
      f16 1024-197-1-88, p=0.0, BiasT=NoneType   |     324.5  |     356.0  |     926.0
      f16 1024-197-1-80, p=0.0, BiasT=NoneType   |     307.9  |     332.0  |     853.0
      f16 1024-197-1-64, p=0.0, BiasT=NoneType   |     204.4  |     235.0  |     720.0
      f16 512-197-1-80, p=0.0, BiasT=NoneType    |     159.6  |     169.0  |     441.0
      f16 32-197-16-80, p=0.0, BiasT=NoneType    |     160.2  |     171.0  |     542.0
      f16 32-197-16-64, p=0.0, BiasT=NoneType    |     108.1  |     123.0  |     462.0
      f16 32-197-16-128, p=0.0, BiasT=NoneType   |     174.3  |     189.0  |     715.0
      f16 256-197-1-88, p=0.0, BiasT=NoneType    |      90.9  |      97.0  |     257.0
      f16 16-197-16-88, p=0.0, BiasT=NoneType    |      91.1  |      98.0  |     314.0
      f16 16-197-16-64, p=0.0, BiasT=NoneType    |      68.6  |      67.0  |     254.0
      f16 16-197-16-128, p=0.0, BiasT=NoneType   |      93.1  |     100.0  |     383.0
      f16 1024-82-8-64, p=0.0, BiasT=NoneType    |     434.2  |     511.0  |    1768.0
      f16 150-256-16-64, p=0.0, BiasT=NoneType   |     507.2  |     581.0  |    1747.0
      f16 64-256-12-64, p=0.0, BiasT=NoneType    |     171.2  |     195.0  |     591.0
      f16 1-4096-16-40, p=0.0, BiasT=NoneType    |     827.6  |     951.0  |    1849.0
      f16 1-16384-16-40, p=0.0, BiasT=NoneType   |   11905.1  |   13613.0  |   28511.0
      f16 1-4096-16-80, p=0.0, BiasT=NoneType    |    1162.4  |    1343.0  |    2010.0
      f16 1-16384-16-80, p=0.0, BiasT=NoneType   |   16914.8  |   20807.0  |   29779.0
      f16 4-4096-16-40, p=0.0, BiasT=NoneType    |    3017.1  |    3443.0  |    7452.0
      f16 4-16384-16-40, p=0.0, BiasT=NoneType   |   46661.8  |   53621.0  |  114900.0
      f16 4-4096-16-80, p=0.0, BiasT=NoneType    |    4329.3  |    5279.0  |    7817.0
      f16 4-16384-16-80, p=0.0, BiasT=NoneType   |   67646.9  |   82607.0  |          
      f16 256-4096-16-64, p=0.0, BiasT=NoneType  |  182128.3  |  206636.0  |          
      f16 8-2048-20-128, p=0.0, BiasT=NoneType   |    2992.3  |    3665.0  |    5728.0
      f16 1-2048-4-128, p=0.0, BiasT=NoneType    |     132.5  |     125.0  |     166.0
      f16 1-2048-8-128, p=0.0, BiasT=NoneType    |     201.7  |     220.0  |     316.0
      f16 1-4096-4-128, p=0.0, BiasT=NoneType    |     396.1  |     434.0  |     581.0
      f16 1-4096-8-128, p=0.0, BiasT=NoneType    |     648.8  |     781.0  |    1096.0
      f16 1-8192-4-128, p=0.0, BiasT=NoneType    |    1271.5  |    1540.0  |    2044.0
      f16 1-8192-8-128, p=0.0, BiasT=NoneType    |    2446.4  |    2879.0  |    4007.0
      f16 2-2048-4-128, p=0.0, BiasT=NoneType    |     202.2  |     221.0  |     349.0
      f16 2-2048-8-128, p=0.0, BiasT=NoneType    |     331.8  |     400.0  |     634.0
      f16 2-4096-4-128, p=0.0, BiasT=NoneType    |     647.2  |     785.0  |    1153.0
      f16 2-4096-8-128, p=0.0, BiasT=NoneType    |    1237.9  |    1453.0  |    2211.0
      f16 2-8192-4-128, p=0.0, BiasT=NoneType    |    2442.0  |    2877.0  |    4084.0
      f16 2-8192-8-128, p=0.0, BiasT=NoneType    |    4646.8  |    5699.0  |    8189.0
      f16 16-128-16-16, p=0.0, BiasT=NoneType    |      67.7  |      66.0  |     125.0
      f16 16-128-16-32, p=0.0, BiasT=NoneType    |      67.6  |      67.0  |     127.0
      f16 16-128-16-64, p=0.0, BiasT=NoneType    |      68.3  |      67.0  |     126.0
      f16 16-128-16-128, p=0.0, BiasT=NoneType   |      68.1  |      66.0  |     138.0
      f16 16-512-16-16, p=0.0, BiasT=NoneType    |     173.1  |     195.0  |     459.0
      f16 16-512-16-32, p=0.0, BiasT=NoneType    |     179.2  |     205.0  |     510.0
      f16 16-512-16-64, p=0.0, BiasT=NoneType    |     206.5  |     234.0  |     595.0
      f16 16-512-16-128, p=0.0, BiasT=NoneType   |     336.3  |     405.0  |     794.0
      f16 16-1024-16-16, p=0.0, BiasT=NoneType   |     652.5  |     743.0  |    1606.0
      f16 16-1024-16-32, p=0.0, BiasT=NoneType   |     662.4  |     749.0  |    1699.0
      f16 16-1024-16-64, p=0.0, BiasT=NoneType   |     764.7  |     862.0  |    2007.0
      f16 16-1024-16-128, p=0.0, BiasT=NoneType  |    1242.6  |    1493.0  |    2395.0
      f16 384-197-1-88, p=0.3, BiasT=NoneType    |     188.0  |     187.0  |     423.0
      f16 384-197-1-64, p=0.3, BiasT=NoneType    |     145.7  |     156.0  |     344.0
      f32 1024-197-1-88, p=0.0, BiasT=NoneType   |    1249.9  |    1196.0  |    1691.0
      f32 1024-197-1-80, p=0.0, BiasT=NoneType   |    1237.6  |    1186.0  |    1628.0
      f16 1024-197-1-64, p=0.3, BiasT=NoneType   |     351.2  |     378.0  |     861.0
      f16 32-197-16-80, p=0.3, BiasT=NoneType    |     239.0  |     236.0  |     616.0
      b16 256-197-1-88, p=0.0, BiasT=NoneType    |      91.0  |      96.0  |     256.0
      f16 16-197-16-88, p=0.3, BiasT=NoneType    |     130.8  |     131.0  |     352.0
      f16 150-256-16-64, p=0.3, BiasT=NoneType   |     913.0  |     961.0  |    2295.0
      f16 1-16384-16-40, p=0.3, BiasT=NoneType   |   23301.9  |   24744.0  |   42698.0
      f32 1-4096-16-80, p=0.0, BiasT=NoneType    |    5099.9  |    5176.0  |    7766.0
      b16 8-2048-20-128, p=0.0, BiasT=NoneType   |    2953.6  |    3537.0  |    5604.0
      b16 1-4096-4-128, p=0.0, BiasT=NoneType    |     396.2  |     428.0  |     568.0
      b16 1-4096-8-128, p=0.0, BiasT=NoneType    |     641.9  |     762.0  |    1068.0
      f32 1-8192-4-128, p=0.0, BiasT=NoneType    |    5726.0  |    6069.0  |    9737.0
      f16 2-2048-8-128, p=0.3, BiasT=NoneType    |     523.9  |     576.0  |     860.0
      b16 2-4096-4-128, p=0.0, BiasT=NoneType    |     642.4  |     764.0  |    1127.0
      b16 2-8192-8-128, p=0.0, BiasT=NoneType    |    4593.1  |    5535.0  |    7932.0
      f32 16-128-16-16, p=0.0, BiasT=NoneType    |      68.0  |      68.0  |     146.0
      b16 16-128-16-128, p=0.0, BiasT=NoneType   |      74.1  |      72.0  |     138.0
      b16 16-1024-16-16, p=0.0, BiasT=NoneType   |     652.8  |     744.0  |    1936.0
      f16 16-1024-16-32, p=0.3, BiasT=NoneType   |    1395.3  |    1453.0  |    2643.0
      b16 16-1024-16-64, p=0.0, BiasT=NoneType   |     756.2  |     851.0  |    1989.0
      b16 16-1024-16-128, p=0.0, BiasT=NoneType  |    1230.8  |    1475.0  |    2403.0

Times are in microseconds (us).

[----- attention (attn_bias=<class 'xformers.ops.fmha.attn_bias.LowerTriangularMask'>) -----]
                                                           |    v3    |    v2    |    eager  
1 threads: ----------------------------------------------------------------------------------
      f16 384-197-1-80, p=0.0, BiasT=LowerTriangularMask   |   102.3  |    99.0  |    97945.0
      f16 32-197-16-128, p=0.0, BiasT=LowerTriangularMask  |   141.2  |   142.0  |   145722.0
      f16 16-197-16-128, p=0.0, BiasT=LowerTriangularMask  |    77.5  |    76.0  |    62703.0
      f16 1-4096-16-40, p=0.0, BiasT=LowerTriangularMask   |   517.5  |   586.0  |  1944397.0
      f16 4-4096-16-40, p=0.0, BiasT=LowerTriangularMask   |  1629.4  |  1860.0  |  7787570.0
      f16 2-8192-4-128, p=0.0, BiasT=LowerTriangularMask   |  1388.6  |  1627.0  |  3897440.0
      f16 16-128-16-64, p=0.0, BiasT=LowerTriangularMask   |    67.8  |    67.0  |     9072.0
      f16 16-512-16-128, p=0.0, BiasT=LowerTriangularMask  |   230.0  |   267.0  |   511932.0

Times are in microseconds (us).

[----------------- attention (attn_bias=<class 'torch.Tensor'>) ----------------]
                                              |     v3    |     v2     |   eager 
1 threads: ----------------------------------------------------------------------
      f16 1-16384-16-80, p=0.0, BiasT=Tensor  |  23930.3  |   26992.0  |  46913.0
      f16 4-4096-16-80, p=0.0, BiasT=Tensor   |   6054.0  |    6803.0  |  12463.0
      f16 4-16384-16-80, p=0.0, BiasT=Tensor  |  95213.0  |  107346.0  |         
      f16 1-2048-8-128, p=0.0, BiasT=Tensor   |    280.4  |     308.0  |    461.0
      f16 2-4096-8-128, p=0.0, BiasT=Tensor   |   1674.0  |    1798.0  |   3154.0
      f16 16-512-16-16, p=0.0, BiasT=Tensor   |    278.8  |     307.0  |    746.0
      f16 16-512-16-32, p=0.0, BiasT=Tensor   |    294.6  |     316.0  |    799.0
      f16 16-512-16-64, p=0.0, BiasT=Tensor   |    312.0  |     342.0  |    914.0

Times are in microseconds (us).
```
</details>

<details>
<summary>BW pass - up to 10x faster thanks to better parallelization</summary>

```
[------------------------ attention backward (attn_bias=<class 'NoneType'>) ------------------------]
                                                                 |     v3     |     v2     |  vanilla
1 threads: ------------------------------------------------------------------------------------------
      f16 384-197-1-88, p=0.0, BiasT=NoneType, BiasGrad=False    |     456.1  |     657.0  |    814.0
      f16 384-197-1-80, p=0.0, BiasT=NoneType, BiasGrad=False    |     437.7  |     633.0  |    759.0
      f16 384-197-1-64, p=0.0, BiasT=NoneType, BiasGrad=False    |     373.1  |     435.0  |    644.0
      f16 1024-197-1-88, p=0.0, BiasT=NoneType, BiasGrad=False   |    1136.5  |    1683.0  |   2087.0
      f16 1024-197-1-80, p=0.0, BiasT=NoneType, BiasGrad=False   |    1089.5  |    1621.0  |   1939.0
      f16 1024-197-1-64, p=0.0, BiasT=NoneType, BiasGrad=False   |     941.0  |     948.0  |   1632.0
      f16 512-197-1-80, p=0.0, BiasT=NoneType, BiasGrad=False    |     552.3  |     812.0  |    985.0
      f16 32-197-16-80, p=0.0, BiasT=NoneType, BiasGrad=False    |     562.5  |     817.0  |   1023.0
      f16 32-197-16-64, p=0.0, BiasT=NoneType, BiasGrad=False    |     501.6  |     495.0  |    874.0
      f16 32-197-16-128, p=0.0, BiasT=NoneType, BiasGrad=False   |     654.0  |     926.0  |   1317.0
      f16 256-197-1-88, p=0.0, BiasT=NoneType, BiasGrad=False    |     336.3  |     477.0  |    570.0
      f16 16-197-16-88, p=0.0, BiasT=NoneType, BiasGrad=False    |     340.8  |     476.0  |    592.0
      f16 16-197-16-64, p=0.0, BiasT=NoneType, BiasGrad=False    |     278.4  |     255.0  |    478.0
      f16 16-197-16-128, p=0.0, BiasT=NoneType, BiasGrad=False   |     375.0  |     520.0  |    707.0
      f16 1024-82-8-64, p=0.0, BiasT=NoneType, BiasGrad=False    |    1610.0  |    1953.0  |   3609.0
      f16 150-256-16-64, p=0.0, BiasT=NoneType, BiasGrad=False   |    2245.9  |    2239.0  |   3823.0
      f16 64-256-12-64, p=0.0, BiasT=NoneType, BiasGrad=False    |     744.3  |     791.0  |   1260.0
      f16 1-4096-16-40, p=0.0, BiasT=NoneType, BiasGrad=False    |    3009.8  |   27160.0  |   3504.0
      f16 1-16384-16-40, p=0.0, BiasT=NoneType, BiasGrad=False   |   43228.1  |  462370.0  |  55770.0
      f16 1-4096-16-80, p=0.0, BiasT=NoneType, BiasGrad=False    |    3835.0  |   27876.0  |   3945.0
      f16 1-16384-16-80, p=0.0, BiasT=NoneType, BiasGrad=False   |   59291.3  |  449408.0  |  57321.0
      f16 4-4096-16-40, p=0.0, BiasT=NoneType, BiasGrad=False    |   10293.2  |   30334.0  |  13620.0
      f16 4-16384-16-40, p=0.0, BiasT=NoneType, BiasGrad=False   |  178613.7  |  492130.0  |         
      f16 4-4096-16-80, p=0.0, BiasT=NoneType, BiasGrad=False    |   14968.9  |   29623.0  |  14323.0
      f16 4-16384-16-80, p=0.0, BiasT=NoneType, BiasGrad=False   |  238149.5  |  476788.0  |         
      f16 256-4096-16-64, p=0.0, BiasT=NoneType, BiasGrad=False  |  630288.8  |  717958.0  |         
      f16 8-2048-20-128, p=0.0, BiasT=NoneType, BiasGrad=False   |   10254.3  |   16890.0  |  10524.0
      f16 1-2048-4-128, p=0.0, BiasT=NoneType, BiasGrad=False    |     407.1  |    6488.0  |    339.0
      f16 1-2048-8-128, p=0.0, BiasT=NoneType, BiasGrad=False    |     822.7  |    6572.0  |    633.0
      f16 1-4096-4-128, p=0.0, BiasT=NoneType, BiasGrad=False    |    1577.0  |   25789.0  |   1110.0
      f16 1-4096-8-128, p=0.0, BiasT=NoneType, BiasGrad=False    |    2391.6  |   28389.0  |   2081.0
      f16 1-8192-4-128, p=0.0, BiasT=NoneType, BiasGrad=False    |    4665.1  |  110260.0  |   3680.0
      f16 1-8192-8-128, p=0.0, BiasT=NoneType, BiasGrad=False    |    7820.8  |  113587.0  |   7122.0
      f16 2-2048-4-128, p=0.0, BiasT=NoneType, BiasGrad=False    |     817.8  |    6582.0  |    640.0
      f16 2-2048-8-128, p=0.0, BiasT=NoneType, BiasGrad=False    |    1254.9  |    7362.0  |   1182.0
      f16 2-4096-4-128, p=0.0, BiasT=NoneType, BiasGrad=False    |    2393.8  |   28313.0  |   2102.0
      f16 2-4096-8-128, p=0.0, BiasT=NoneType, BiasGrad=False    |    4016.8  |   29191.0  |   4170.0
      f16 2-8192-4-128, p=0.0, BiasT=NoneType, BiasGrad=False    |    7835.5  |  113347.0  |   7147.0
      f16 2-8192-8-128, p=0.0, BiasT=NoneType, BiasGrad=False    |   15580.3  |  116299.0  |  14296.0
      f16 16-128-16-16, p=0.0, BiasT=NoneType, BiasGrad=False    |     223.9  |     245.0  |    291.0
      f16 16-128-16-32, p=0.0, BiasT=NoneType, BiasGrad=False    |     220.5  |     245.0  |    292.0
      f16 16-128-16-64, p=0.0, BiasT=NoneType, BiasGrad=False    |     222.3  |     244.0  |    288.0
      f16 16-128-16-128, p=0.0, BiasT=NoneType, BiasGrad=False   |     222.0  |     244.0  |    286.0
      f16 16-512-16-16, p=0.0, BiasT=NoneType, BiasGrad=False    |     671.5  |     675.0  |    981.0
      f16 16-512-16-32, p=0.0, BiasT=NoneType, BiasGrad=False    |     714.2  |     748.0  |   1088.0
      f16 16-512-16-64, p=0.0, BiasT=NoneType, BiasGrad=False    |     826.4  |     927.0  |   1274.0
      f16 16-512-16-128, p=0.0, BiasT=NoneType, BiasGrad=False   |    1348.3  |    1884.0  |   1699.0
      f16 16-1024-16-16, p=0.0, BiasT=NoneType, BiasGrad=False   |    2312.8  |    2657.0  |   3524.0
      f16 16-1024-16-32, p=0.0, BiasT=NoneType, BiasGrad=False   |    2423.7  |    2866.0  |   3722.0
      f16 16-1024-16-64, p=0.0, BiasT=NoneType, BiasGrad=False   |    2763.5  |    3404.0  |   4242.0
      f16 16-1024-16-128, p=0.0, BiasT=NoneType, BiasGrad=False  |    4385.5  |    6802.0  |   5083.0
      f16 384-197-1-88, p=0.3, BiasT=NoneType, BiasGrad=False    |     661.5  |     919.0  |    859.0
…